Update patch set 1
Patch Set 1:

(1 comment)

Code looks good to me but I have a question.  

Also, check out this patch in neutron:  https://review.openstack.org/#/c/164079/

This changes the name of the midonet_task_state to midonet_data_state in preparation for the data sync feature (some field name changes too).  I think it makes sense to have this patch to match the spec that we had (so keep it as is) and provide future patches for the table change separately as part of data sync work.  But considering the timing, if we want them to be functional together, it would make sense to change it now.  Will leave it up to you to decide on how you want to organize your patches.

Patch-set: 1
Label: Verified=0