Update patch set 1
Patch Set 1: Code-Review-2

How many times do we need to repeat the same mistake? ZOOM cannot handle a Scala Object because of some deserialisation issue. It can handle a pure POJO, so if you really want to do this, you need to define C3POState in Java, that is, if you still want to write Java.

Patch-set: 1
Label: Code-Review=-2