VxGW: Publish hosts to VTEP service logical switch
The VtepController will create a logical switch in which it will publish
dummy remote mac locations associated to each MidoNet Agent host using
the IP dedicated for tunnels with the VTEP.  The purpose is that VTEPs
are able to recognize remote endpoints before traffic comes from them.
For some VTEPs, this is required as otherwise they will drop the ingress
traffic on unknown tunnels.  This case hits MidoNet, as we may have
hosts (e.g., gateway, flooding proxy..) that tunnel traffic to the VTEP
without having any VMs on it.  This implies that no mac remote entries
will have been pushed to the remote tables of the VTEP by the time that
the gateway sends traffic through the tunnel.

Ref.: MNA-145

Change-Id: I94253f48e9561996a7d5e436c4360d00da241b2b
Signed-off-by: Galo Navarro <galo@midokura.com>