Update patch set 4
Patch Set 4:

(5 comments)

There are C3PO translation logic that I'm not familiar with (for example, I don't know how router/bridge ports are linked).  I'll organize it so that each header is a task object with sub-sections for the task actions (CREATE, UPDATE, etc).  Hopefully that makes it more clear.

Patch-set: 4