Migrate to new Vtep Data models
Remove the old ovsdb-dependent vtep-related
data model and make midonet-brain and midonet-api
depend on the ovsdb-agnostic data models in the
cluster module.

This is part of the preparation for the
introduction of a new vtep support implementation
based on the standard ovsdb library, instead of
our own forked ovsdb branch.

Ref: MN-2543

Change-Id: I698e1b047c672d29b1b3f189de2fa49be629f809
Signed-off-by: Ernest Artiaga <ernest@midokura.com>