Update patch set 1
Patch Set 1:

This option protects writes into the GC's card table (which happen whenever there's a write) with an if, to check if that address range has already been marked. This avoids cache line ping ponging in concurrent applications. I ran the perf tests with this on but saw no differences, since we're still dominated by other costs. However, we'll have a lot of objects tenured (packet contexts, managed flows, etc) to which a lot of threads will be writing, so intuitively I think this  should be on (and Doug Lea recommends it as well). More here: https://blogs.oracle.com/dave/entry/false_sharing_induced_by_card

Patch-set: 1