Run simulation before dhcp handling
The old behaviour of dhcp handling occurring before simulation meant
that if flow tracing was enabled for the dhcp flow, it would never hit
the rule and therefore never trace anything. To solve this, we run
simulation before dhcp handling, even though in the dhcp case we throw
the simulation result away.

Ref MN-3265

Change-Id: I09c1a575e461677353a7e224c76b42b012d2099f
Signed-off-by: Ivan Kelly <ivan@midokura.com>