Update patch set 1
Patch Set 1: Code-Review-2

For DHCP, running a simulation is calling handleDHCP. If we run a faux simulation, then whatever we trace is pretty much useless, and gives only to noise. To really fix this we need to move DHCP handling into the simulation like we did with BGP (see RoutingWorkflow).

Patch-set: 1