Update patch set 1
Patch Set 1:

I've been looking at how to do this within the simulation itself. It's not as simple as moving this into simulation.Bridge, as the DHCPImpl needs a handle to a DataClient, which is not available in Bridge, nor BridgeManager, nor VirtualTopologyActor. I could add it all the way down, but it seems like a dirty solution. A cleaner solution would be to have a dhcp actor, which the dhcpimpl could do all the dhcp for the agent, and to which the bridge simulation could punt dhcp requests too when needed. However, this also doesn't seem like a great idea, since we are moving away from akka.

Patch-set: 1