Update patch set 1
Patch Set 1: -Code-Review

We can brainstorm about what's the best way to implement this. I'm not sure how this stuff gets created and bootstrapped and what are the underlying assumptions :/ We may have to create a DhcpBridgePort or something. It does feel like something we'd probably want to solve in the new cluster bits to avoid duplicated effort. 

What I'm most worried with this patch is tracing useless things. Maybe we can check whether tracing is enabled, and in that case skip the faux simulation?

Patch-set: 1
Label: Code-Review=0