New conf: provide close method
Adds a new close method to ObservableConf that will close the
underlying zookeeper NodeCache.

Also, because it's a common pattern to just ask for a configuration
source for a single-call use. Add and onClose method that takes
a closure and ensures the ObservableConf is thrown away after the
enclosed block runs.

This patch addes uses of this in MidoNodeConfigurator and ConfMinion
to remove ZK watcher leaks.

Change-Id: I4ce3c35e8a0d00ae00d264ed3f461ccb5f80ac27
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>