Add support for maps to Storage/Zoom.
Adds support for CreateMap, DeleteMap, AddMapEntry, UpdateMapEntry,
and DeleteMapEntry to ZookeeperObjectMapper.multi(). Does not add
support to InMemoryStorage, or support for reading map entries, or
single-operation methods (e.g. ZookeeperObjectMapper.createMap()), as
these are not currently necessary. It's likely that at some point we
will want read/observable methods for the agent to use, but for the
time being they will continue using the old DataClient library.

Signed-off-by: Brandon Berg <bberg@midokura.com>
Change-Id: Ic55169b69d1309218a2a158c1f5a09fe870c62dc