Update patch set 1
Patch Set 1:

I've added a check in Bridge, so that if a packet is found to be dhcp, simulation is aborted. The check should be very fast in most cases (one long comparison).

Checking if tracing is enabled, as suggested by duarte, doesn't work for cases, for example if two trace requests want to capture this package.

Patch-set: 1
Label: Code-Review=0