Update patch set 1
Patch Set 1: Code-Review+2

Hi Tomohiko. This is going to break the logic of the TopologyZoomUpdater, but it should not be difficult to fix. Please, go ahead with the merge, and I'll fix the updater in a following patch.

Patch-set: 1
Reviewer: Ernest Artiaga <1002750@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+2