Add per agent trace rule limits
When a trace request is enabled, it inserts a rule into a device
chain. This change adds a limit in the number of flows it will match
before it disabled itself. When it disables itself it will not remove
itself from the chain, but it will not trigger any more traces. Each
time the trace is enabled or disabled, the rule is reset.

Ref MN-3197

Change-Id: I702a582bf1c63cb7baa7108c0b700a11d220f27e
Signed-off-by: Ivan Kelly <ivan@midokura.com>