Always reset origMatch after fork action
Previously, if any fork action threw an exception (such as NotYet) the
origMatch would not be reset back to its original value. This caused the
match to be incorrect the next time that packet context ran.

In the case that brought this to my attention a
IndexOutOfBoundsException was being thrown when a vlan id was being
popped, after it had already been popped in a previous attempt then then
postponed. This was causing flake on the l2gw mdts tests.

Ref MNA-175

Change-Id: If3c7bd4cf7f7dca2c3483278849fd26ff25ab6e4
Signed-off-by: Ivan Kelly <ivan@midokura.com>