Update patch set 3
Patch Set 3: Code-Review-1

Thanks Tomohiko:

I think the best solution to allowing the cluster to perform operations on objects using exclusive ownership is to modify the ZOOM transaction manager such that ownership-agnostic ops (Create, Update, Delete) do not fail. In this case, it involves removing a simple check for those operations. Alternatively, we could introduce a new Update op, that ignores the ownership, or simply modifying only the current Update.

The reason for this, is that ownership is indicative of state. In the case of hosts, having an owner indicates that the host is alive. Therefore, only the agents should be allowed to take ownership of a Host object. Since the ownership is exclusive, only one agent can take ownership of a Host at a time (when let's say we start two agent instances on the same host).

If we allow the cluster to update the ownership, then the cluster would actually represent the host alive state. If you remember from the ownership patch, when we update the owner, we delete and recreate the ephemeral node ZK, to make sure that it corresponds to the client session that performed the last update. Hence, the ownership state would change when the cluster node would close its ZK connection, rather than when the agent does it.

To summarize, I think the best solution is to modify the ZOOM TransactionManager to:
- either update the Create, Update, Delete (or perhaps only the Update) ops to ignore the ownership, or;
- introduce new op(s) that should allow use modifying an object, even if it has exclusive ownership. Maybe these ops could be exposed only via the StorageWithOwnership trait (e.g. ForcedUpdate or something like this).

Anyway the solution to this is simply bypassing the check that throws the exception in the validateOwner() method in TransactionManager.

Patch-set: 3
Reviewer: Alex Bikfalvi <1002751@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1