Update patch set 2
Patch Set 2: Code-Review-1

(3 comments)

It looks good to me, I'm putting -1 because it seems to me that testing that a delete on an ownerless object with exclusive ownership is missing. According to the commit message, a delete for such an object should work if not owner is specified in the delete call isn't it?

Patch-set: 2
Label: Code-Review=-1