Bootstrap: Datapath init is part of bootstrap
This patch takes datapath initialization out of the DatapathController
and moves it earlier in the bootstrap sequence. This means components
don't have to depend on the DatapathController's DatapathReady
message.

It also simplifies datapath initialization so that we delete a
pre-existing datapath instead of trying to recover it by deleting all
the ports and flows.

Ref MN-3407

Change-Id: Id202fab545fb5fb6448493c1577d3a04102712c6
Signed-off-by: Duarte Nunes <duarte@midokura.com>