Add node support to Storage/Zoom
Adds support in ZookeeperObjectMapper.multi() for CreateNode,
UpdateNode, and DeleteNode operations. These allow the creation of
nodes at arbitrary locations, which is needed to allow the topology
translators to write topology changes and replicated map changes
atomically.

Does not add support to InMemoryStorage, single-operation methods
(e.g. ZookeeperObjectMapper.createNode()), or support for reading
nodes, as these are not currently needed. The ability to subscribe to
updates may be required at some point, but for the time being the
agent will continue to access replicated maps via the old DataClient
interface.

Change-Id: I48712b4a5307ca551b323862e11f85694a2048de
Signed-off-by: Brandon Berg <bberg@midokura.com>