Compat API initial check in
This commit adds a basic Cluster service that uses an embedded Jetty to
provide a REST API aiming to serve the same endpoints as the existing
MidoNet API, but built on the new storage stack.  The ultimate goal is
to allow passing all MDTS tests (which use the REST API), while not
using the old MidoNet API (that requires the old storage stack).

Change-Id: I88f5d5ebad7404734298e9ee3136b3653d44742c
Signed-off-by: Galo Navarro <galo@midokura.com>
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>