Update patch set 10
Patch Set 10: Code-Review-1

(13 comments)

Thanks Alex for taking care of this. It looks mostly good to me. There's a missing test in VirtualTopologyRedirectorTest and a few things need to be added so that the api can also obtain pools and that API tests encompass added proto buffs.

Patch-set: 10
Label: Code-Review=-1