Output: Adaptively increase buffer size
This patch catches BufferOverflowExceptions in the FlowProcessor and
PacketExecutor classes, which respectively write out flows and packets
to the datapath, so as to adaptively increase the buffer size. On each
exception, we double the buffer size, which starts at 64k and grows to
a maximum of 4MB.

A BufferOverflowException may happen if a packet is received from a
network with an MTU over the buffer size or if we're executing a
packet or creating a flow and the size of the serialized flow actions
exceed the buffer size.

Change-Id: I437231bcbef1c41897420f441ebd027769098435
Signed-off-by: Duarte Nunes <duarte@midokura.com>