Update patch set 1
Patch Set 1:

I created this for the earlier discussion. I was planning to include this to my patches for the new InterfaceScanner but it's not happening today and I just create a separate patch only for this.

Patch-set: 1
Reviewer: Taku Fukushima <1002743@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0