Move common API code to brain
This patch moves all the code from the REST API that is not coupled to
the old storage layer into the Brain module.  This will enable a new
implementation of the Resource classes that goes against the new storage
stack.

There are some cleanups of unnecessary code, and minor refactors to
decouple some odd classes from the old storage models in order to move
them safely to brain.

Also, removes Cloudstack Auth, as it's unused.

Change-Id: I2405206e4df5e36101d9bf318b84ed6b13dab49d
Signed-off-by: Galo Navarro <galo@midokura.com>