Move common API code to brain
This patch moves all the code from the REST API that is not coupled to
the old storage layer into the Brain module.  This will enable a new
implementation of the Resource classes that goes against the new storage
stack.

There are some cleanups of unnecessary code, and minor refactors to
decouple some odd classes from the old storage models in order to move
them safely to brain.

Also, removes Cloudstack Auth, as it's unused.

Also, removes a lot of useless, redundant Javadoc.

Change-Id: I5dc3a723d292836fd2e69267199a93d7e1d70cb4
Signed-off-by: Galo Navarro <galo@midokura.com>