Floating IP Translator sets a Prov Router route.
Floating IP Translator needs to update (add / delete) a route to the
Provider Router upon assignment / unassignment of a Floating IP to a
VIF Port.

An integration test is to come in a separate patch, which cannot pass at
the moment because of the bug that a creation of Router Interfce Port
doesn't link Router and Network. (MN-216)

SNAT / DNAT rules are not handled yet.

Fix MNA-185
Change-Id: I01b49bd0990417898cce8535e0a1d767a7763ee2