Add a Router Interface on R-Interface Port CREATE.
Add a peer Router Port for the created Router Interface (Network) Port,
attach it to the Router and link it to the two ports. Also add a route
to the subnet as well as a route to the Meta Data Service if a DHCP has
been configured with the Network.

When the user adds a Router Interface port, which connects a Network to
a Router, Neutron plugin (on the Neutron side) first makes a CREATE PORT
call with device owner ROUTER INTERFACE, and subsequently makes another
call to ADD-ROUTER-INTERFACE. C3PO doesn't receive any resource CRUD
task corresponding to this, therefore it needs to perform the operation
right after the creation of Router Interface Port.

Provides a new integration test where it drives the creation of a
Network, a tenant Router, and a Router Interface port, and tests if the
corresponding Network and Routers are correctly linked.

Fix MNA-216
Change-Id: If060fe7060c4b097d4c137b2e66a19c5ac5f1be4