Edge router uplink
Add support for uplink networks edge routers. An uplink network has no
corresponding Midonet network. An edge router is a router which is
connected to an uplink network in the Neutron topology. We model this
in the Midonet topology by creating a port on the edge router that is
bound to an interface on a host.

Signed-off-by: Brandon Berg <bberg@midokura.com>
Change-Id: I28fa8ddbcf79171982ea3f7037daeb854a89b936