Update patch set 7
Patch Set 7: Code-Review+1

+1 just because no one else has looked at it.

I think the takeaway lesson from this should be that if we find ourselves searching through what could potentially be thousands of Zookeeper nodes, we need to find a better way. Even if we didn't have the Neutron data available, we could have solved this by adding the gateway port ID to the Midonet router model. This is exactly why we have backreferences, and there's no reason we can't add more if we need them.

Patch-set: 7
Label: Code-Review=+1