Sim: Flow deletion is now thread safe
Due to the FlowController being sharded across the simulation threads,
the assumption that only one thread performs flow deletions no
longer holds.

Since the NetlinkRequestBroker is not thread safe for multiple
publishing threads, this patch allows the FlowProcessor to hold
multiple Brokers, each with its channel through which writes and
deletions are performed.

Note that because flow operations are synchronous, it will still be
the case that flow creation and deletion are ordered by their
sequence number.

Ref MNA-263

Change-Id: I8d55db4a075e74548de88638b9787cb12d7f53e2
Signed-off-by: Duarte Nunes <duarte@midokura.com>