Update patch set 1
Patch Set 1:

Several months ago I tried to fix the problems with importing the Gradle file, but it turned out that the IDEA Gradle plugin was based on a different version than the version of Gradle we were using for the official build, so I gave up on that and just decided to build the project on the command line. It only takes a couple of seconds to rebuild, and IDEA will automatically prompt you to reload it, so there's really no reason not to do this. I was reluctant at first, but overall I find it less of a hassle than refreshing the Gradle project from within IDEA.

Patch-set: 1