Implement DHCP for new storage stack
This is a simple implementation that replicates the current
functionality by fetching the DHCP configuration from ZOOM.  The
implementation is blocking, as the current one, which is something worth
improving but out of scope for this issue.

The PacketWorkflow is now adapted to remain agnostic of the backend
choice, and instead rely on a factory that provides the right class to
access DHCP configuration.

Fixes.: MNA-259
Change-Id: I67f7d101fb96ae24f840255d95c8702a69670a79
Signed-off-by: Galo Navarro <galo@midokura.com>