Make perfCompile Gradle config extend provided
For some reason, adding a Gradle dependency configuration to an IDEA
scope will also add its ancestor configurations to that IDEA scope. So
if perfCompile extends compile, adding it to IDEA's PROVIDED scope
will cause all the compile and runtime dependencies to be added to the
IDEA module as Provided dependencies, rather than Runtime or Compile
dependencies. I'm making perfCompile extend provided to address this
issue.

Change-Id: I5735d753f8852d14d11184b4d44b701a8139881e
Signed-off-by: Brandon Berg <bberg@midokura.com>