Remove flush from Storage interface
This change removes the FLUSH feature from the Storage interface
implemented by ZOOM, as it's not the right layer of abstraction for the
feature.  In the new storage stack, the FLUSH operation is specific to
the Importer (codename c3po) who will react to a FLUSH Neutron task by
re-importing all tasks necessary to recreate the Neutron storage into
MidoNet. To achieve this, it will of course interact with ZOOM as a
client-side library to the ZooKeeper storage, but performing the
necessary actions & semantics for the FLUSH belongs in the Importer's
layer as a ZOOM client. Other users of ZOOM (e.g., the Agent) do not
need to be aware of the existence of a Flush operation.

The reason for removing this now is to simplify the Storage interface to
enable other implementations different than the ZOOM storage (e.g., a
Cassandra backend).

The sequence of actions to be taken on a FLUSH will be detailed in a
separate spec, but briefly will be the following:

- ZOOM will expose a currVersion node stored in ZK. This node will
  contain a version number. This version number will determine the root
  path for the zoom storage. For example, given a currVersion of 1, when
  a ZOOM instance is initialized it'll point itself at root path
  /midonet/zoom/1.
- ZOOM exposes an Observable that notifies about changes in the
  currVersion node.
- When the importer sees a FLUSH operation, it will ask ZOOM to write a
  currVersion + 1 into the currVersion node.
- All services using ZOOM (e.g., an agent) will be notified about a
  change in the currVersion node. The users of ZOOM should react to this
  by closing the ZOOM instance and reopening a new one. As per the
  second point above, ZOOM will now point at /midonet/zoom/2

This works both in the Agent, Importer, Topology API and other
components. All of them will restart immediately, and their new ZOOM
instance will point directly at the new root path. The importer will
start importing new TASK entries, and clients will gradually
re-subscribe to the desired entities at their new locations.

More complex solution could allow transparently swapping the underlying
observables.

As noted, all these changes will be specced and detailed in a separate
document.

Change-Id: I4c883fb2c3ce5264caa35d3f4cc8aaa2de8da286
Signed-off-by: Galo Navarro <galo@midokura.com>