Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

Thank you for preparing this, Galo. In general Removing flush from the API at this moment is a good idea, and will simplify API dependencies. I wonder, however, if we can keep the flush implementation as is with ZookeeperObjectMapper for now. Furthermore, your comment says that the version node should be watched by clients of ZOOM and not ZOOM itself. Does that mean that ZOOM will NOT closes class and instance subscriptions upon a version bump?

Patch-set: 1
Label: Code-Review=-1