SystemD: kill by control-group rather than process
Our systemD service was configured to kill by process, which means that
only the top level process would be manually killed. We have code to
shutdown subprocesses in the process, but this gets timed out if the
shutdown hasn't occurred in 200ms. Upstart kills by process group, so
this hasn't been a problem in the past. This patch updates our systemD
configuration to do the same.

Ref MN-3385

Change-Id: I636d4bebe0c71a3e8c36e7b903c86ce19aafb60e
Signed-off-by: Ivan Kelly <ivan@midokura.com>