Remove Topology.Host.PortBinding message
Remove the Host.PortBinding message class and replace the Host
message's port_bindings field with port_ids. This allows us to use
Zoom Bindings to automatically update the Host message when deleting a
bound Port or updating a Port's host_id field, which simplifies
translation and solves a problem arising from the fact that Neutron
doesn't provide us with a delete Port or delete PortBinding task when
deleting a Network that has a DHCP port.

The PortBinding message only contained redundant information already
present in the corresponding Port message, so no information is lost.

Change-Id: I042fb69794f51a39dd786117341e3db58784da25