Test the last processed in a separate eventually.
Test the last processed task ID in separate eventually blocks than tests
for checking the translation results have properly been persisted. This
is because C3POMinion updates the last processed task ID table here in a
separate call to MySQL DB that's made after the translation of a Neutron
task transaction is completed. Therefore, we need to test it in a separate
eventually block.

Fix MN-3427

Change-Id: If65450047fcb8fdfa41d76360b9a2f4e881284a2