Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

This has a bit of a chicken and egg problem since the conf system requires.. zk. I think this can be solved by just providing a conf file and not forcing this cluster node to connect to zk to fetch the embedded zk conf. Need to check with Guillermo.

Patch-set: 1