Update patch set 90
Patch Set 90:

(4 comments)

Still reviewing. I get the impression that it would be worth splitting this in three patches, one with the mock client, another with the udpates to the models, schema etc. in brain/vtep, and another with the final adaptation. I bet it'll be easier to get the extra reviewer that way :)

Patch-set: 90