mn-conf: support cross-layer refs
Validation of config key values was failing when key
refs across layers existed in the user-provided configuration,
due to the fact that the config object use for validation was
not backed by the schema.

Also, the -t option in the set command was broken since 'local'
became the default for the -h option, as -h would take precedence
over -t and would shadow it.

Change-Id: If7cc7305b994232a3e40a0086db2f702f9b00cfc
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>