Update patch set 4
Patch Set 4:

Hi Tomohiko, please review the latest version. After looking a bit into the failing tests that you also pointed out, I removed a small section that was relying on ZOOM as I think it was just implicitly testing the flush by making a cleanup task within a test. 

In the current approach, the tests clean zk and restore the backend on every run of every test, as common practise. This allows us not to expose any dangerous apis on the backend to wipe the storage. 

Regarding the code for restoring subscriptions, is it ok to keeping it in version control instead of commented out? We could also move it to a design document btw.

Patch-set: 4
Label: Code-Review=0