Cache the write version
The write version should be cached and updated by a watcher rather than
read every time as is currently the case. JsonVersionZkSerializer checks
the write version before serializing or deserializing. As most of our
reads and writes to zookeeper use this serializer, and the write version
read call is a synchronous call, this effectively doubles the time taken
to make any zookeeper call.

With this patch, write version is only read once, the first time it is
accessed, and at this point a watcher is put on the node. Any subsequent
change to the value in zookeeper will result in it being reread. This is
effectively a reimplementation of Curator's node cache. I didn't want to
have to create a new curator client just for this though.

Ref MN-2950

Change-Id: Iacd4766c73a307e8dfde1ea59b1e34c37e71070b
Signed-off-by: Ivan Kelly <ivan@midokura.com>