Only create ZK structure if it doesn't exist
Each agent tries to create the midonet zookeeper structure on both. The
previous implementation did this eagerly, ignoring NodeExists errors if
the node already existed, which it usually did. This was done to avoid
two roundtrips to zookeeper for each element of the path. However, when
a create error at the zookeeper server, an ErrorTxn is added to the
zookeeper log, so the client still pays the penalty of the write. The
most common case here is that the structure already exists, so we are
paying the penalty for 50 useless writes on every boot. Furthermore, if
multiple nodes in the cluster are being rebooted at the same time, this
will result in a herd of writes to the zookeeper cluster, possibly
interfering with the correct booting of the cluster.

In this patch, the problem is solved by checking for the existence of
the node before trying to create. This does add an extra roundtrip for
the exists() call, but in the vast vast majority of cases this is all
that's needed.

Ref MN-2950

Change-Id: I630ff05c2f93d02f99ed47c10a27a0651b209246
Signed-off-by: Ivan Kelly <ivan@midokura.com>