The HostMapper emits only distinct hosts
This patch adds tests to verify that the HostMapper only emits a single
host notification when either the host_id field in a port is cleared or
when the host_id is removed from the list of hosts of a tunnel zone.
Also, this patch tests that non-modifying updates to the host, tunnel
zone, or port does not result in a host update.

Ref: MNA-306

Change-Id: I223f9d0e6c874b0155393dd6e4dd8e377dec583d
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>