VTA updates topology cache before notifying subscribers
Previously, whenever a topology device was updated, the VTA would
first notify all pending subscribers of the updated, and then
update the internal topology cache. Due to this latency, subscribers
that would receive a notification (either a device update or the
completion of the future returned by the NotYetException) will
find that a subsequent call to tryAsk may not return the updated
device or will again throw a NotYetException.

This patch addresses this problem, by updating the VTA cache before
notifying subscribers of any device update.

This also fixes some flaky tests, such as the
VirtualTopologyRedirectorTest, which assumed the VTA cache is up-to-
date, following a notification.

Fixes: MNA-274

Change-Id: Ib85d75a16c51dd0ee3d0f4a9f956e42c7919e319
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>