OVSDB Vtep schema translations
This patch is part of the port of Ovsdb support
to the new Ovsdb client libraries. It consolidates
the code to translate from internal OVSDB-based
vtep data schemas to implementation-agnostic
vtep data models.

These schemas provide low-level support and are
based on the specs in
http://openvswitch.org/docs/vtep.5.pdf
and the reference implementation of the ovsdb
library: https://github.com/opendaylight/ovsdb

Ref: MN-2543

Change-Id: I4f17f55a48fa981a25b04e5e3fb33caa1ad88aba
Signed-off-by: Ernest Artiaga <ernest@midokura.com>