Low-level Ovsdb VTEP operations
This patch is part of the port of Ovsdb support
to the new Ovsdb client libraries. It provides
wrappers for common boilerplate code used to
access the low-level Ovsdb operations via the
Ovsdb client.

Note that no checks are done at this level: the
classes using these tools are responsible for
passing sensible values (according to their needs)
into the methods.

Ref: MN-2543

Change-Id: Ia9bdaa67f676247e7298e98ba513af08dc28cd83
Signed-off-by: Ernest Artiaga <ernest@midokura.com>