MDTS: script to bootstrap new stack
This patch provides a script that is able to take a MDTS system
running after the boot script, and tweak it to use the compat
api and the new storage stack.  For this, it brings up an new
container with the cluster, and tells agents to use the new stack.

Change-Id: I7f971016851c6b0ffcf45216b6f8d379e5851324
Fixes: MNA-287
Signed-off-by: Galo Navarro <galo@midokura.com>