Read tunnel zone members in parallel
When the agent boots, it subscribes to the tunnel zone, and reads all
the members in the tunnel zone. Previously the hosts in the tunnel zone
were read one by one, which meant that the time to read all members was
proportional to the number of hosts in the cluster.

This patch reads all the tunnel zone members in parallel, so that we
only take the latency hit of a single zookeeper read.

Ref MN-2950

Change-Id: Ifd38c8dc9fa6788c4e18367985d21466aeb512ae
Signed-off-by: Ivan Kelly <ivan@midokura.com>