Introduce MidoSequencers
This is a simple class that abstracts access and management to
the various centralized counters that MidoNet maintains in the system.
These are so far two: agent tunnel keys and VNIs for Vxlan Gateways.

The Sequencer provides a "current" method, that yields the current value
of the counter, as well as a "next" method that reserves the next
number in the sequence.

The implementation is based on ZK, using Curator's SharedCounter.  There
are no usages yet, but this will be introduced in the Neutron Importer
and the Compat API to generate tunnel keys and VNIs in the new storage
stack.

Ref.: MNA-309
Signed-off-by: Galo Navarro <galo@midokura.com>

Change-Id: If277b236ecc1e11e314844deba3ed121071a9464