Make Directory friendly to rx Observers
Adds a wrapper for DirectoryCallback that turns it into an Observable. I
also got rid of DirectoryCallback.Add and Directory.Void as they would
have needed their own wrappers and didn't add anything useful to the
code. I also added a method to the reactor so that observables can be
observed on a reactor.

Ref MN-2950

Change-Id: If541b9e01f711f79c302681dd16017989e608d21
Signed-off-by: Ivan Kelly <ivan@midokura.com>