Make port activation asynchronous
Is an agent has a large number of ports, at boot time, using synchronous
write, each one will occupy the reactor thread for an entire
roundtrip. This means the last port to be activated will be delayed by
<num_ports> x <write_latency>. This patch changes the write path for
activation to use async writes so that we don't hit this latency penalty.

This change is only for the pre-zoom
implementation. ZoomTransactionManager uses curator transactions which
uses multi which isn't available in async form before Zookeeper
3.5.0-alpha.

Ref MN-2950

Change-Id: Icf11839e819fa165542dd17007c01ddb2239de54
Signed-off-by: Ivan Kelly <ivan@midokura.com>