Compat API: Resource updates copies subresources
Currently, when an URI resource mapped to a ZOOM object is updated
we simply store the ZOOM object in storage as is. The failure to
include the old subresources corresponding to this object means
the resource after update will lose all its subresources.

This patch fixes this error, by ensuring that updated resources
copy the subresources from the old ones.

Fixes: MNA-331

Change-Id: Ia9d8f7e9b6082e017d82ab17a79ed597e11e5fb3
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>