Fixing Flaky Mapper Tests
This patch fixes flaky tests in:

-PortGroupMapperTest:
The flaky test was "The mapper emits new device on port group update".
In the PortGroupMapper, the distinctUntilChanged now appears after the
Zoom conversion. This is because the protobuff has a field that does not
appear in the simulation object. Consequently, two proto buffs may be
different although the simulation objects are the same.

-PoolMapperTest:
The flaky tests were: Inactive pool member added, Inactive pool member
updates to active, and Inactive pool member updates to disabled.

To resolve these flaky tests, this patch does two things:
-It include the members of a Pool in the equality test. This way if an
inactive member gets added or updated, the PoolMapper emits an update.
-Comparing the members, active members, and disabled members is now done
regardless of order in the equals method.

Ref: MN-3479

Change-Id: Ie9332314ccac3b60fb15601b65c229fa4f638156
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>