Fixing Flaky PoolMapper Tests
The flaky tests were: Inactive pool member added, Inactive pool member
updates to active, and Inactive pool member updates to disabled.

To resolve these flaky tests, this patch does two things:
-It include the members of a Pool in the equality test, as well as the
loadBalancerId, and the healthMonitorId. This way if any of these
fields change, the PoolMapper emits an update. Additionally, testing
that members of two load-balancers are the same is performed using
the sameElements method instead of ==, which is reference equality
in the case of an object of class Array.
-The PoolMapper builds a pool simulation object with members that
appear in the same order as in the protocol buffer.

Ref: MN-3479

Change-Id: Ie9332314ccac3b60fb15601b65c229fa4f638156
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>