Fixing a potential NPE in LoadBalancer
This patch fixes a potential NPE that could occur in method equals of
class LoadBalancer when comparing the array of VIPs. This will not
happen in the current code because the LoadBalancerMapper never builds a
LoadBalancer with a null array of vips. It is however a good idea to not
play with fire.

Change-Id: I890561d42491820cd05f8690c563ff6772d38dbb
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>